+++ src/main/java/com/petspace/dev/domain/Favorite.java
@@ -1,5 +1,6 @@
 package com.petspace.dev.domain;
 
+import com.petspace.dev.domain.user.User;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
