+++ src/main/java/com/petspace/dev/domain/Room.java
@@ -1,6 +1,7 @@
 package com.petspace.dev.domain;
 
 import com.petspace.dev.domain.image.RoomImage;
+import com.petspace.dev.domain.user.User;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
