+++ src/main/java/com/petspace/dev/domain/user/Role.java
@@ -1,4 +1,4 @@
-package com.petspace.dev.domain;
+package com.petspace.dev.domain.user;
 
 import lombok.AllArgsConstructor;
