+++ src/main/java/com/petspace/dev/domain/user/oauth/OauthAttributes.java
@@ -1,8 +1,8 @@
-package com.petspace.dev.oauth;
+package com.petspace.dev.domain.user.oauth;
 
-import com.petspace.dev.domain.Role;
+import com.petspace.dev.domain.user.Role;
 import com.petspace.dev.domain.Status;
-import com.petspace.dev.domain.User;
+import com.petspace.dev.domain.user.User;
 
 import java.util.Arrays;
 import java.util.Map;
