+++ src/main/java/com/petspace/dev/dto/user/UserJoinRequestDto.java
@@ -1,9 +1,9 @@
 package com.petspace.dev.dto.user;
 
-import com.petspace.dev.domain.OauthProvider;
-import com.petspace.dev.domain.Role;
+import com.petspace.dev.domain.user.OauthProvider;
+import com.petspace.dev.domain.user.Role;
 import com.petspace.dev.domain.Status;
-import com.petspace.dev.domain.User;
+import com.petspace.dev.domain.user.User;
 import lombok.Getter;
 
 import javax.validation.constraints.NotBlank;
