+++ src/main/java/com/petspace/dev/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.petspace.dev.repository;
 
-import com.petspace.dev.domain.User;
+import com.petspace.dev.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
